COLORADO-1628 Fix the breadcrumb path calculation to remove blanks
The current breadcrumb calculation takes the first <dc:title> tag in the DC
datastream.  The sorting order of the <dc:title> tags (in the case of more than
one tag) has empty tags appearing as the first items to be selected by the
query.  This fix alters the query to change the sort order to make empty tags
the last items to be selected and thus, if a non-empty tag exists it will be
selected as the name of the breadcrumb.

There is also a problem with the breadcrumb query processing that will attempt
to process failed queries.  The test for failure was incorrect and allowed some
failed queries to be treated as successes.  This resulted in recursively
calling the breadcrumb processing an additional 10 times and failing each time.
This resulted in 10 '>' symbols instead of an error message.  This fix tests
for falure and puts up an error message in the breadcrumb trail.

The breadcrumb calculation should be rewritten to be more robust.