ISLANDORA-655 OwnerId being wiped out.
Looks like the "modifyObject" SOAP call requires an owner ID...
By default, we were providing NULL, which would result in it being
blanked out.  I've made it so that it grabs the owner ID from the
object profile when it is NULL, so that it should be maintained
by default.