ISLANDORA-764: Rewrite core permissions handling
Needed to be able to give permissions from other places, like XACML...
In order to do so, we have to break out of the context of Drupal
permissions, and the use of user_access().

Instead, we have introduced the function fedora_repository_check_perm()
(in addition to deprecating the old fedora_repository_access(), to avoid
possible confusion with the hook_access() used by Drupal nodes). This
new function calls a hook of the same name (hook_fedora_repository_check_perm()),
to allow other modules to influence the outcome.

The default/core implementation should function as permissions previously
worked.