Follow File Name Conventions
File naming conventions now state that all PHP files (except templates) should
be lower-case and use underscores to separate words in a filename.

includes/IslandoraTuque.inc, and includes/IslandoraTuqueWrapper.inc are not
included in this commit as there is another pull request that corrects their
names in place, and we would like to avoid merge conflicts.