Fixed bug in audio/video solution pack
We were implicitly passing the extra datastream version paramerter
in the islandora video solution pack because we were going to the
url: islandora/object/pid/datastream/MP3/view/name.mp3 and because
the datastream view fucntion now took an extra parameter this was
getting implicitly passed on. So we would try to download version
name.mp3. This makes sure that it won't happen.