Update advanced-search.md (#1860)
The word decedent means one who has died. Based on the context of the documentation the word you are looking for is descendant. I changed it in the documentation above. The only pratical implication is that I also changed the name of the filter in the documentation to field_descendant_of.