Added line about COMPOSE_PROJECT_NAME (#2006)
I think we should encourage people to change the COMPOSE_PROJECT_NAME variable when setting up a new install. If they don't change it now, it's harder to change later if they decide to run multiple instances. Also, by leaving it set to isle-dc we increase the risk of accidentally overwriting the existing containers when spinning up another set of isle-dc containers.